Revert "Restructuring for group b"
