--- group_b/cleaned_data/top_words_for_100.csv
@@ -1,101 +1,0 @@
-,0,1,2,3,4,5,6,7,8,9
-0,land,end,avanova,woman,large,top,sleeve,shirt,short,womens
-1,dress,casual,pocket,sleeve,long,woman,loose,tunic,with,short
-2,black,shirt,large,men,the,size,adult,cap,white,logo
-3,earring,piercing,ring,ear,steel,nose,stud,jewelry,hoop,gauge
-4,watch,men,quartz,waterproof,wrist,steel,analog,dial,stainless,band
-5,woman,top,size,plus,shirt,womens,large,black,sleeve,tank
-6,shirt,top,blouse,sleeve,casual,neck,woman,tunic,short,loose
-7,keychain,key,chain,keyring,ring,gift,car,for,holder,lanyard
-8,shoe,kid,sneaker,toddler,boy,girl,little,baby,led,light
-9,swimsuit,piece,one,swimwear,bathing,suit,tankini,woman,monokini,swim
-10,pant,jean,waist,denim,high,woman,leg,short,elastic,casual
-11,cardigan,sweater,open,front,long,knit,sleeve,woman,coat,womens
-12,silver,sterling,necklace,pendant,925,charm,jewelry,bracelet,chain,heart
-13,hat,cap,baseball,beanie,adjustable,sun,men,brim,for,black
-14,costume,hat,cosplay,christmas,hoodie,disney,adult,black,anime,sweatshirt
-15,men,shirt,sleeve,large,black,fit,slim,long,short,tee
-16,swing,dress,casual,loose,pocket,woman,sleeve,summer,shirt,tunic
-17,cocktail,dress,party,woman,lace,vintage,evening,bodycon,floral,sleeveless
-18,ballet,dance,leotard,skirt,girl,tutu,gymnastics,dress,for,tulle
-19,apron,for,black,blue,and,large,kitchen,with,size,unisex
-20,sleepwear,pajama,set,sleep,short,nightgown,sleeve,woman,loungewear,ekouaer
-21,pullover,sweater,long,sleeve,top,knit,neck,woman,loose,casual
-22,charm,necklace,pendant,bracelet,jewelry,bead,crystal,silver,for,pandora
-23,girl,denim,jean,skirt,size,blue,short,pack,legging,cotton
-24,boot,ankle,heel,sandal,toe,shoe,woman,sock,platform,high
-25,racerback,tank,woman,sport,top,bra,workout,short,shirt,black
-26,sock,men,crew,boot,pair,pack,size,shoe,glove,for
-27,jersey,shirt,football,men,youth,large,fan,team,sport,ncaa
-28,watch,leather,band,strap,black,22mm,20mm,genuine,buckle,men
-29,dress,neck,ruffle,sleeve,woman,mini,tie,short,casual,long
-30,glass,reading,reader,eyeglass,computer,blocking,light,frame,hinge,blue
-31,scarf,shawl,wrap,winter,soft,silk,infinity,knit,cashmere,warm
-32,woman,dress,apron,vintage,sleeve,print,black,neck,large,womens
-33,swing,dress,vintage,woman,party,floral,cocktail,sleeve,neck,1950s
-34,earring,dangle,drop,for,stud,woman,hoop,girl,jewelry,tassel
-35,boy,kid,little,shirt,youth,big,toddler,set,short,girl
-36,watch,quartz,dial,wrist,steel,analog,woman,stainless,lady,gold
-37,girl,toddler,kid,little,dress,baby,princess,pink,shoe,mary
-38,waterproof,jacket,men,outdoor,rain,hiking,winter,lightweight,snow,windproof
-39,belt,buckle,leather,men,waist,for,genuine,ratchet,black,jean
-40,polo,lauren,ralph,men,shirt,pony,fit,horse,classic,custom
-41,jumpsuit,romper,woman,leg,wide,pant,long,piece,sleeveless,short
-42,lens,sunglass,frame,polarized,glass,eyeglass,oakley,replacement,black,mirror
-43,gym,workout,short,men,running,athletic,dry,training,pocket,bodybuilding
-44,top,shirt,blouse,tank,woman,loose,sleeve,casual,sleeveless,neck
-45,woman,pajama,womens,robe,set,short,cotton,sleeve,soft,large
-46,dress,woman,bodycon,sexy,midi,skirt,black,lace,sleeveless,mini
-47,necklace,pendant,jewelry,for,gold,woman,silver,gift,chain,crystal
-48,workout,yoga,running,short,sport,bra,gym,woman,top,athletic
-49,mask,face,reusable,for,halloween,filter,adult,washable,masquerade,with
-50,shoe,sock,sneaker,shoelace,pair,lace,black,size,no,for
-51,dress,bridesmaid,wedding,gown,prom,evening,formal,party,lace,long
-52,sweatshirt,hoodie,men,pullover,hooded,jacket,hoodies,sweater,zip,3d
-53,dress,woman,sleeve,neck,long,maxi,black,large,bodycon,short
-54,kid,shoe,boy,toddler,water,slipper,girl,slip,little,beach
-55,sunglass,frame,lens,polarized,fashion,woman,retro,oversized,glass,eye
-56,hat,kid,boy,baby,toddler,girl,winter,cap,cotton,beanie
-57,bag,purse,handbag,wallet,leather,shoulder,crossbody,tote,woman,clutch
-58,goggles,kid,swim,protection,fog,swimming,anti,uv,boy,for
-59,tie,cufflink,pin,necktie,men,set,silk,bow,gift,clip
-60,earring,silver,sterling,stud,gold,925,hoop,plated,dangle,14k
-61,cotton,men,shirt,100,pack,fit,short,black,large,sleeve
-62,bracelet,bangle,bead,for,cuff,jewelry,adjustable,charm,leather,woman
-63,sunglass,polarized,men,driving,frame,sport,glass,lens,for,protection
-64,dress,party,floral,woman,vintage,sleeve,print,lace,line,long
-65,bikini,swimsuit,bathing,suit,swimwear,piece,set,up,woman,high
-66,nike,black,white,adidas,men,air,size,jordan,shoe,dri
-67,top,neck,sleeve,woman,blouse,shirt,long,crop,tunic,romwe
-68,michael,kor,leather,black,satchel,jet,tote,large,bag,medium
-69,men,running,jogger,pant,compression,sport,athletic,sweatpants,sock,workout
-70,sweatshirt,pullover,hoodie,long,sleeve,hoodies,top,woman,casual,hooded
-71,slipper,slip,indoor,house,shoe,sandal,on,slide,outdoor,foam
-72,fitbit,band,for,silicone,replacement,versa,strap,charge,with,compatible
-73,bag,travel,tote,backpack,luggage,crossbody,shoulder,canvas,shopping,for
-74,black,woman,boot,heather,large,size,grey,girl,shirt,the
-75,woman,waist,pant,high,panty,black,size,skirt,womens,stretch
-76,men,boxer,short,brief,underwear,pant,fit,cargo,pack,jean
-77,band,watch,compatible,series,apple,40mm,iwatch,38mm,42mm,44mm
-78,coat,jacket,winter,warm,woman,fur,fleece,faux,hooded,outwear
-79,costume,halloween,cosplay,dress,party,adult,for,girl,outfit,woman
-80,necklace,choker,chain,jewelry,pearl,pendant,for,woman,gold,statement
-81,floral,flower,woman,rose,print,pink,top,size,sleeve,dress
-82,baby,girl,toddler,outfit,clothes,month,set,boy,pant,romper
-83,necklace,chain,steel,stainless,pendant,gold,jewelry,ring,silver,inch
-84,pencil,dress,bodycon,work,woman,business,sleeve,office,wear,midi
-85,bra,push,padded,woman,strapless,up,sport,seamless,underwire,bralette
-86,shirt,men,button,sleeve,down,casual,fit,short,long,slim
-87,wine,red,dress,sleeve,woman,neck,long,top,casual,shirt
-88,yoga,pant,legging,workout,high,waist,tummy,control,woman,pocket
-89,shirt,tee,graphic,top,sleeve,funny,short,woman,letter,print
-90,umbrella,windproof,folding,compact,travel,rain,patio,auto,close,outdoor
-91,gift,for,birthday,bracelet,necklace,jewelry,mother,mom,girl,day
-92,dress,summer,maxi,floral,casual,woman,long,sleeveless,sleeve,print
-93,beach,dress,summer,woman,casual,maxi,floral,sundress,sleeveless,cover
-94,ring,wedding,band,silver,gold,size,for,engagement,woman,sterling
-95,face,gaiter,bandana,scarf,mask,balaclava,neck,cover,for,dust
-96,dress,floral,floerns,woman,ruffle,print,sleeve,summer,neck,short
-97,glove,black,leather,size,strap,with,cap,pack,large,buckle
-98,face,reusable,washable,mask,mouth,dust,filter,cover,breathable,adjustable
-99,silver,ring,sterling,925,gold,plated,zirconia,size,cubic,cz
